Update SD preparation section
Fixed an issue in SD preparation caused by the files.zip containing the files for tiramisu in the root instead of the `/wiiu/` folder. I'm not sure for this change but I know this caused a lot of headache when I was following the guide and I'm sure it probably has for other people. Hence why I made this fix.